+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Not released
 
 * Add ``scalib.tools.ContextExecutor``, as a solution to ``LookupError`` in
   ``get_config()``.
+* Fix numerical underflow in ``BPState`` when multiple traces are used.
 
 v0.5.3 (2023/02/21)
 -------------------